[ANDROSDK-1964] Move ConstantFields to network package
